+++ packages/fx-core/tests/component/coordinator/coordinator.create.test.ts
@@ -1,6 +1,6 @@
 import "mocha";
 
-import { err, Inputs, ok, Platform, UserError } from "@microsoft/teamsfx-api";
+import { err, Inputs, ok, Platform, SystemError, UserError } from "@microsoft/teamsfx-api";
 import { assert } from "chai";
 import fs from "fs-extra";
 import * as sinon from "sinon";
@@ -22,6 +22,8 @@ import {
   ScratchOptions,
 } from "../../../src/question/create";
 import { MockTools, randomAppName } from "../../core/utils";
+import { MockedUserInteraction } from "../../plugins/solution/util";
+import { OfficeAddinGenerator } from "../../../src/component/generator/officeAddin/generator";
 
 const V3Version = MetadataV3.projectVersion;
 describe("coordinator create", () => {
@@ -545,5 +547,93 @@ describe("coordinator create", () => {
 
     assert.isTrue(res2.isOk());
     assert.equal(generator.args[0][2], TemplateNames.Tab);
+  });
+});
+
+describe("Office Addin", async () => {
+  const sandbox = sinon.createSandbox();
+  const tools = new MockTools();
+  tools.ui = new MockedUserInteraction();
+  setTools(tools);
+
+  beforeEach(() => {
+    sandbox.stub(fs, "ensureDir").resolves();
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  it("should scaffold taskpane successfully", async () => {
+    const v3ctx = createContextV3();
+    v3ctx.userInteraction = new MockedUserInteraction();
+
+    sandbox.stub(OfficeAddinGenerator, "generate").resolves(ok(undefined));
+    sandbox
+      .stub(settingsUtil, "readSettings")
+      .resolves(ok({ trackingId: "mockId", version: V3Version }));
+    sandbox.stub(settingsUtil, "writeSettings").resolves(ok(""));
+
+    const inputs: Inputs = {
+      platform: Platform.VSCode,
+      folder: ".",
+      [QuestionNames.ProjectType]: ProjectTypeOptions.outlookAddin().id,
+      [QuestionNames.AppName]: randomAppName(),
+      [QuestionNames.Scratch]: ScratchOptions.yes().id,
+    };
+    const res = await coordinator.create(v3ctx, inputs);
+    assert.isTrue(res.isOk());
+  });
+
+  it("should return error if app name is invalid", async () => {
+    const v3ctx = createContextV3();
+    v3ctx.userInteraction = new MockedUserInteraction();
+    const inputs: Inputs = {
+      platform: Platform.VSCode,
+      folder: ".",
+      [QuestionNames.AppName]: "__invalid__",
+      [QuestionNames.Scratch]: ScratchOptions.yes().id,
+      [QuestionNames.ProjectType]: ProjectTypeOptions.outlookAddin().id,
+    };
+
+    const res = await coordinator.create(v3ctx, inputs);
+    assert.isTrue(res.isErr() && res.error instanceof InputValidationError);
+  });
+
+  it("should return error if app name is undefined", async () => {
+    const v3ctx = createContextV3();
+    v3ctx.userInteraction = new MockedUserInteraction();
+    const inputs: Inputs = {
+      platform: Platform.VSCode,
+      folder: ".",
+      [QuestionNames.AppName]: undefined,
+      [QuestionNames.Scratch]: ScratchOptions.yes().id,
+      [QuestionNames.ProjectType]: ProjectTypeOptions.outlookAddin().id,
+    };
+
+    const res = await coordinator.create(v3ctx, inputs);
+    assert.isTrue(res.isErr() && res.error instanceof MissingRequiredInputError);
+  });
+
+  it("should return error if OfficeAddinGenerator returns error", async () => {
+    const v3ctx = createContextV3();
+    v3ctx.userInteraction = new MockedUserInteraction();
+
+    const mockedError = new SystemError("mockedSource", "mockedError", "mockedMessage");
+    sandbox.stub(OfficeAddinGenerator, "generate").resolves(err(mockedError));
+    sandbox
+      .stub(settingsUtil, "readSettings")
+      .resolves(ok({ trackingId: "mockId", version: V3Version }));
+    sandbox.stub(settingsUtil, "writeSettings").resolves(ok(""));
+
+    const inputs: Inputs = {
+      platform: Platform.VSCode,
+      folder: ".",
+      [QuestionNames.AppName]: randomAppName(),
+      [QuestionNames.Scratch]: ScratchOptions.yes().id,
+      [QuestionNames.ProjectType]: ProjectTypeOptions.outlookAddin().id,
+    };
+    const res = await coordinator.create(v3ctx, inputs);
+    assert.isTrue(res.isErr() && res.error.name === "mockedError");
   });
 });
